Move sub-classes of "PropertyGrid"  classes to their own files
