+++ src/System.Windows.Forms/src/System/Windows/Forms/PropertyGrid.SelectedObjectConverter.cs
@@ -1,0 +1,20 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+#nullable disable
+
+using System.ComponentModel;
+
+namespace System.Windows.Forms
+{
+    public partial class PropertyGrid
+    {
+        internal class SelectedObjectConverter : ReferenceConverter
+        {
+            public SelectedObjectConverter() : base(typeof(IComponent))
+            {
+            }
+        }
+    }
+}
