+++ src/System.Windows.Forms/src/System/Windows/Forms/PropertyGrid.cs
@@ -12,7 +12,6 @@ using System.Drawing;
 using System.Drawing.Design;
 using System.Drawing.Imaging;
 using System.Globalization;
-using System.IO;
 using System.Reflection;
 using System.Runtime.InteropServices;
 using System.Windows.Forms.ComponentModel.Com2Interop;
@@ -25,7 +24,7 @@ namespace System.Windows.Forms
 {
     [Designer("System.Windows.Forms.Design.PropertyGridDesigner, " + AssemblyRef.SystemDesign)]
     [SRDescription(nameof(SR.DescriptionPropertyGrid))]
-    public class PropertyGrid : ContainerControl, IComPropertyBrowser, Ole32.IPropertyNotifySink
+    public partial class PropertyGrid : ContainerControl, IComPropertyBrowser, Ole32.IPropertyNotifySink
     {
         private readonly DocComment _doccomment;
         private int _dcSizeRatio = -1;
@@ -5067,705 +5066,6 @@ namespace System.Windows.Forms
             }
 
             base.WndProc(ref m);
-        }
-
-        internal abstract class SnappableControl : Control
-        {
-            protected PropertyGrid ownerGrid;
-            internal bool userSized;
-
-            public abstract int GetOptimalHeight(int width);
-            public abstract int SnapHeightRequest(int request);
-
-            public SnappableControl(PropertyGrid ownerGrid)
-            {
-                this.ownerGrid = ownerGrid;
-                SetStyle(ControlStyles.OptimizedDoubleBuffer, true);
-            }
-
-            public override Cursor Cursor
-            {
-                get
-                {
-                    return Cursors.Default;
-                }
-                set => base.Cursor = value;
-            }
-
-            protected override void OnControlAdded(ControlEventArgs ce)
-            {
-            }
-
-            public Color BorderColor { get; set; } = SystemColors.ControlDark;
-
-            protected override void OnPaint(PaintEventArgs e)
-            {
-                base.OnPaint(e);
-                Rectangle r = ClientRectangle;
-                r.Width--;
-                r.Height--;
-
-                using var borderPen = BorderColor.GetCachedPenScope();
-                e.Graphics.DrawRectangle(borderPen, r);
-            }
-        }
-
-        public class PropertyTabCollection : ICollection
-        {
-            private readonly PropertyGrid _owner;
-
-            internal PropertyTabCollection(PropertyGrid owner)
-            {
-                _owner = owner;
-            }
-
-            /// <summary>
-            ///  Retrieves the number of member attributes.
-            /// </summary>
-            public int Count
-            {
-                get
-                {
-                    if (_owner is null)
-                    {
-                        return 0;
-                    }
-                    return _owner._viewTabs.Length;
-                }
-            }
-
-            object ICollection.SyncRoot
-            {
-                get
-                {
-                    return this;
-                }
-            }
-
-            bool ICollection.IsSynchronized
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            /// <summary>
-            ///  Retrieves the member attribute with the specified index.
-            /// </summary>
-            public PropertyTab this[int index]
-            {
-                get
-                {
-                    if (_owner is null)
-                    {
-                        throw new InvalidOperationException(SR.PropertyGridPropertyTabCollectionReadOnly);
-                    }
-                    return _owner._viewTabs[index];
-                }
-            }
-
-            public void AddTabType(Type propertyTabType)
-            {
-                if (_owner is null)
-                {
-                    throw new InvalidOperationException(SR.PropertyGridPropertyTabCollectionReadOnly);
-                }
-                _owner.AddTab(propertyTabType, PropertyTabScope.Global);
-            }
-
-            public void AddTabType(Type propertyTabType, PropertyTabScope tabScope)
-            {
-                if (_owner is null)
-                {
-                    throw new InvalidOperationException(SR.PropertyGridPropertyTabCollectionReadOnly);
-                }
-                _owner.AddTab(propertyTabType, tabScope);
-            }
-
-            /// <summary>
-            ///  Clears the tabs of the given scope or smaller.
-            ///  tabScope must be PropertyTabScope.Component or PropertyTabScope.Document.
-            /// </summary>
-            public void Clear(PropertyTabScope tabScope)
-            {
-                if (_owner is null)
-                {
-                    throw new InvalidOperationException(SR.PropertyGridPropertyTabCollectionReadOnly);
-                }
-                _owner.ClearTabs(tabScope);
-            }
-
-            void ICollection.CopyTo(Array dest, int index)
-            {
-                if (_owner is null)
-                {
-                    return;
-                }
-                if (_owner._viewTabs.Length > 0)
-                {
-                    System.Array.Copy(_owner._viewTabs, 0, dest, index, _owner._viewTabs.Length);
-                }
-            }
-            /// <summary>
-            ///  Creates and retrieves a new enumerator for this collection.
-            /// </summary>
-            public IEnumerator GetEnumerator()
-            {
-                if (_owner is null)
-                {
-                    return Array.Empty<PropertyTab>().GetEnumerator();
-                }
-
-                return _owner._viewTabs.GetEnumerator();
-            }
-
-            public void RemoveTabType(Type propertyTabType)
-            {
-                if (_owner is null)
-                {
-                    throw new InvalidOperationException(SR.PropertyGridPropertyTabCollectionReadOnly);
-                }
-                _owner.RemoveTab(propertyTabType);
-            }
-        }
-
-        internal class SelectedObjectConverter : ReferenceConverter
-        {
-            public SelectedObjectConverter() : base(typeof(IComponent))
-            {
-            }
-        }
-
-        private class PropertyGridServiceProvider : IServiceProvider
-        {
-            private readonly PropertyGrid _owner;
-
-            public PropertyGridServiceProvider(PropertyGrid owner)
-            {
-                _owner = owner;
-            }
-
-            public object GetService(Type serviceType)
-            {
-                object s = null;
-
-                if (_owner.ActiveDesigner != null)
-                {
-                    s = _owner.ActiveDesigner.GetService(serviceType);
-                }
-
-                if (s is null)
-                {
-                    s = _owner._gridView.GetService(serviceType);
-                }
-
-                if (s is null && _owner.Site != null)
-                {
-                    s = _owner.Site.GetService(serviceType);
-                }
-                return s;
-            }
-        }
-
-        /// <summary>
-        ///  Helper class to support rendering text using either GDI or GDI+.
-        /// </summary>
-        internal static class MeasureTextHelper
-        {
-            public static SizeF MeasureText(PropertyGrid owner, Graphics g, string text, Font font)
-            {
-                return MeasureTextSimple(owner, g, text, font, new SizeF(0, 0));
-            }
-
-            public static SizeF MeasureText(PropertyGrid owner, Graphics g, string text, Font font, int width)
-            {
-                return MeasureText(owner, g, text, font, new SizeF(width, 999999));
-            }
-
-            public static SizeF MeasureTextSimple(PropertyGrid owner, Graphics g, string text, Font font, SizeF size)
-            {
-                SizeF bindingSize;
-                if (owner.UseCompatibleTextRendering)
-                {
-                    bindingSize = g.MeasureString(text, font, size);
-                }
-                else
-                {
-                    bindingSize = (SizeF)TextRenderer.MeasureText(g, text, font, Size.Ceiling(size), GetTextRendererFlags());
-                }
-
-                return bindingSize;
-            }
-
-            public static SizeF MeasureText(PropertyGrid owner, Graphics g, string text, Font font, SizeF size)
-            {
-                SizeF bindingSize;
-                if (owner.UseCompatibleTextRendering)
-                {
-                    bindingSize = g.MeasureString(text, font, size);
-                }
-                else
-                {
-                    TextFormatFlags flags =
-                        GetTextRendererFlags() |
-                        TextFormatFlags.LeftAndRightPadding |
-                        TextFormatFlags.WordBreak |
-                        TextFormatFlags.NoFullWidthCharacterBreak;
-
-                    bindingSize = (SizeF)TextRenderer.MeasureText(g, text, font, Size.Ceiling(size), flags);
-                }
-
-                return bindingSize;
-            }
-
-            public static TextFormatFlags GetTextRendererFlags()
-            {
-                return TextFormatFlags.PreserveGraphicsClipping |
-                        TextFormatFlags.PreserveGraphicsTranslateTransform;
-            }
-        }
-    }
-
-    internal static class AutomationMessages
-    {
-        internal const int PGM_GETBUTTONCOUNT = (int)User32.WM.USER + 0x50;
-        internal const int PGM_GETBUTTONSTATE = (int)User32.WM.USER + 0x52;
-        internal const int PGM_SETBUTTONSTATE = (int)User32.WM.USER + 0x51;
-        internal const int PGM_GETBUTTONTEXT = (int)User32.WM.USER + 0x53;
-        internal const int PGM_GETBUTTONTOOLTIPTEXT = (int)User32.WM.USER + 0x54;
-        internal const int PGM_GETROWCOORDS = (int)User32.WM.USER + 0x55;
-        internal const int PGM_GETVISIBLEROWCOUNT = (int)User32.WM.USER + 0x56;
-        internal const int PGM_GETSELECTEDROW = (int)User32.WM.USER + 0x57;
-        internal const int PGM_SETSELECTEDTAB = (int)User32.WM.USER + 0x58; // DO NOT CHANGE THIS : VC uses it!
-        internal const int PGM_GETTESTINGINFO = (int)User32.WM.USER + 0x59;
-
-        /// <summary>
-        ///  Writes the specified text into a temporary file of the form %TEMP%\"Maui.[file id].log", where
-        ///  'file id' is a unique id that is return by this method.
-        ///  This is to support MAUI interaction with the PropertyGrid control and MAUI should remove the
-        ///  file after used.
-        /// </summary>
-        public static IntPtr WriteAutomationText(string text)
-        {
-            IntPtr fileId = IntPtr.Zero;
-            string fullFileName = GenerateLogFileName(ref fileId);
-
-            if (fullFileName != null)
-            {
-                try
-                {
-                    FileStream fs = new FileStream(fullFileName, FileMode.Create, FileAccess.Write);
-                    StreamWriter sw = new StreamWriter(fs);
-                    sw.WriteLine(text);
-                    sw.Dispose();
-                    fs.Dispose();
-                }
-                catch
-                {
-                    fileId = IntPtr.Zero;
-                }
-            }
-
-            return fileId;
-        }
-
-        /// <summary>
-        ///  Writes the contents of a test file as text.  This file needs to have the following naming convention:
-        ///  %TEMP%\"Maui.[file id].log", where 'file id' is a unique id sent to this window.
-        ///  This is to support MAUI interaction with the PropertyGrid control and MAUI should create/delete this file.
-        /// </summary>
-        public static string ReadAutomationText(IntPtr fileId)
-        {
-            Debug.Assert(fileId != IntPtr.Zero, "Invalid file Id");
-
-            string text = null;
-
-            if (fileId != IntPtr.Zero)
-            {
-                string fullFileName = GenerateLogFileName(ref fileId);
-                Debug.Assert(File.Exists(fullFileName), "Automation log file does not exist");
-
-                if (File.Exists(fullFileName))
-                {
-                    try
-                    {
-                        FileStream fs = new FileStream(fullFileName, FileMode.Open, FileAccess.Read);
-                        StreamReader sr = new StreamReader(fs);
-                        text = sr.ReadToEnd();
-                        sr.Dispose();
-                        fs.Dispose();
-                    }
-                    catch
-                    {
-                        text = null;
-                    }
-                }
-            }
-
-            return text;
-        }
-
-        /// <summary>
-        ///  Generate log file from id.
-        /// </summary>
-        private static string GenerateLogFileName(ref IntPtr fileId)
-        {
-            string fullFileName = null;
-
-            string filePath = System.Environment.GetEnvironmentVariable("TEMP");
-            Debug.Assert(filePath != null, "Could not get value of the TEMP environment variable");
-
-            if (filePath != null)
-            {
-                if (fileId == IntPtr.Zero) // Create id
-                {
-                    Random rnd = new Random(DateTime.Now.Millisecond);
-                    fileId = new IntPtr(rnd.Next());
-                }
-
-                fullFileName = filePath + "\\Maui" + fileId + ".log";
-            }
-
-            return fullFileName;
-        }
-    }
-
-    /// <summary>
-    ///  Represents the PropertyGrid accessibility object.
-    ///  Is used only in Accessibility Improvements of level3 to show correct accessible hierarchy.
-    /// </summary>
-    internal class PropertyGridAccessibleObject : Control.ControlAccessibleObject
-    {
-        private readonly PropertyGrid _owningPropertyGrid;
-
-        /// <summary>
-        ///  Initializes new instance of PropertyGridAccessibleObject
-        /// </summary>
-        /// <param name="owningPropertyGrid">The PropertyGrid owning control.</param>
-        public PropertyGridAccessibleObject(PropertyGrid owningPropertyGrid) : base(owningPropertyGrid)
-        {
-            _owningPropertyGrid = owningPropertyGrid;
-        }
-
-        /// <summary>
-        ///  Return the child element at the specified point, if one exists,
-        ///  otherwise return this element if the point is on this element,
-        ///  otherwise return null.
-        /// </summary>
-        /// <param name="x">x coordinate of point to check</param>
-        /// <param name="y">y coordinate of point to check</param>
-        /// <returns>Return the child element at the specified point, if one exists,
-        ///  otherwise return this element if the point is on this element,
-        ///  otherwise return null.
-        /// </returns>
-        internal override UiaCore.IRawElementProviderFragment ElementProviderFromPoint(double x, double y)
-        {
-            if (!_owningPropertyGrid.IsHandleCreated)
-            {
-                return null;
-            }
-
-            Point clientPoint = _owningPropertyGrid.PointToClient(new Point((int)x, (int)y));
-
-            Control element = _owningPropertyGrid.GetElementFromPoint(clientPoint);
-            if (element != null)
-            {
-                return element.AccessibilityObject;
-            }
-
-            return base.ElementProviderFromPoint(x, y);
-        }
-
-        /// <summary>
-        ///  Request to return the element in the specified direction.
-        /// </summary>
-        /// <param name="direction">Indicates the direction in which to navigate.</param>
-        /// <returns>Returns the element in the specified direction.</returns>
-        internal override UiaCore.IRawElementProviderFragment FragmentNavigate(UiaCore.NavigateDirection direction)
-        {
-            switch (direction)
-            {
-                case UiaCore.NavigateDirection.Parent:
-                    return null;
-                case UiaCore.NavigateDirection.FirstChild:
-                    return GetChildFragment(0);
-                case UiaCore.NavigateDirection.LastChild:
-                    var childFragmentCount = GetChildFragmentCount();
-                    if (childFragmentCount > 0)
-                    {
-                        return GetChildFragment(childFragmentCount - 1);
-                    }
-                    break;
-            }
-
-            return base.FragmentNavigate(direction);
-        }
-
-        /// <summary>
-        ///  Request to return the element in the specified direction regarding the provided child element.
-        /// </summary>
-        /// <param name="childFragment">The child element regarding which the target element is searched.</param>
-        /// <param name="direction">Indicates the direction in which to navigate.</param>
-        /// <returns>Returns the element in the specified direction.</returns>
-        internal UiaCore.IRawElementProviderFragment ChildFragmentNavigate(AccessibleObject childFragment, UiaCore.NavigateDirection direction)
-        {
-            switch (direction)
-            {
-                case UiaCore.NavigateDirection.Parent:
-                    return this;
-                case UiaCore.NavigateDirection.NextSibling:
-                    int fragmentCount = GetChildFragmentCount();
-                    int childFragmentIndex = GetChildFragmentIndex(childFragment);
-                    int nextChildFragmentIndex = childFragmentIndex + 1;
-                    if (fragmentCount > nextChildFragmentIndex)
-                    {
-                        return GetChildFragment(nextChildFragmentIndex);
-                    }
-
-                    return null;
-                case UiaCore.NavigateDirection.PreviousSibling:
-                    fragmentCount = GetChildFragmentCount();
-                    childFragmentIndex = GetChildFragmentIndex(childFragment);
-                    if (childFragmentIndex > 0)
-                    {
-                        return GetChildFragment(childFragmentIndex - 1);
-                    }
-
-                    return null;
-            }
-
-            return null;
-        }
-
-        /// <summary>
-        ///  Return the element that is the root node of this fragment of UI.
-        /// </summary>
-        internal override UiaCore.IRawElementProviderFragmentRoot FragmentRoot
-        {
-            get
-            {
-                return this;
-            }
-        }
-
-        /// <summary>
-        ///  Gets the accessible child corresponding to the specified index.
-        /// </summary>
-        /// <param name="index">The child index.</param>
-        /// <returns>The accessible child.</returns>
-        internal AccessibleObject GetChildFragment(int index)
-        {
-            if (index < 0)
-            {
-                return null;
-            }
-
-            if (_owningPropertyGrid.ToolbarVisible)
-            {
-                if (index == 0)
-                {
-                    return _owningPropertyGrid.ToolbarAccessibleObject;
-                }
-
-                index--;
-            }
-
-            if (_owningPropertyGrid.GridViewVisible)
-            {
-                if (index == 0)
-                {
-                    return _owningPropertyGrid.GridViewAccessibleObject;
-                }
-
-                index--;
-            }
-
-            if (_owningPropertyGrid.CommandsVisible)
-            {
-                if (index == 0)
-                {
-                    return _owningPropertyGrid.HotCommandsAccessibleObject;
-                }
-
-                index--;
-            }
-
-            if (_owningPropertyGrid.HelpVisible)
-            {
-                if (index == 0)
-                {
-                    return _owningPropertyGrid.HelpAccessibleObject;
-                }
-            }
-
-            return null;
-        }
-
-        /// <summary>
-        ///  Gets the number of children belonging to an accessible object.
-        /// </summary>
-        /// <returns>The number of children.</returns>
-        internal int GetChildFragmentCount()
-        {
-            int childCount = 0;
-
-            if (_owningPropertyGrid.ToolbarVisible)
-            {
-                childCount++;
-            }
-
-            if (_owningPropertyGrid.GridViewVisible)
-            {
-                childCount++;
-            }
-
-            if (_owningPropertyGrid.CommandsVisible)
-            {
-                childCount++;
-            }
-
-            if (_owningPropertyGrid.HelpVisible)
-            {
-                childCount++;
-            }
-
-            return childCount;
-        }
-
-        /// <summary>
-        ///  Return the element in this fragment which has the keyboard focus,
-        /// </summary>
-        /// <returns>Return the element in this fragment which has the keyboard focus,
-        ///  if any; otherwise return null.</returns>
-        internal override UiaCore.IRawElementProviderFragment GetFocus()
-        {
-            return GetFocused();
-        }
-
-        /// <summary>
-        ///  Gets the child control index.
-        /// </summary>
-        /// <param name="controlAccessibleObject">The control accessible object which index should be found.</param>
-        /// <returns>The child accessible index or -1 if not found.</returns>
-        internal int GetChildFragmentIndex(AccessibleObject controlAccessibleObject)
-        {
-            int childFragmentCount = GetChildFragmentCount();
-            for (int i = 0; i < childFragmentCount; i++)
-            {
-                AccessibleObject childFragment = GetChildFragment(i);
-                if (childFragment == controlAccessibleObject)
-                {
-                    return i;
-                }
-            }
-
-            return -1;
-        }
-
-        internal override object GetPropertyValue(UiaCore.UIA propertyID) =>
-            propertyID switch
-            {
-                UiaCore.UIA.NamePropertyId => Name,
-                _ => base.GetPropertyValue(propertyID),
-            };
-    }
-
-    /// <summary>
-    ///  Represents the PropertyGrid inner ToolStrip control.
-    ///  Is used starting with Accessibility Improvements of level 3.
-    /// </summary>
-    internal class PropertyGridToolStrip : ToolStrip
-    {
-        private readonly PropertyGrid _parentPropertyGrid;
-
-        /// <summary>
-        ///  Initializes new instance of PropertyGridToolStrip control.
-        /// </summary>
-        /// <param name="parentPropertyGrid">The parent PropertyGrid control.</param>
-        public PropertyGridToolStrip(PropertyGrid parentPropertyGrid)
-        {
-            _parentPropertyGrid = parentPropertyGrid;
-        }
-
-        /// <summary>
-        ///  Indicates whether or not the control supports UIA Providers via
-        ///  IRawElementProviderFragment/IRawElementProviderFragmentRoot interfaces.
-        /// </summary>
-        internal override bool SupportsUiaProviders => true;
-
-        /// <summary>
-        ///  Constructs the new instance of the accessibility object for this control.
-        /// </summary>
-        /// <returns>The accessibility object for this control.</returns>
-        protected override AccessibleObject CreateAccessibilityInstance()
-        {
-            return new PropertyGridToolStripAccessibleObject(this, _parentPropertyGrid);
-        }
-    }
-
-    /// <summary>
-    ///  Represents the PropertyGridToolStrip control accessibility object.
-    /// </summary>
-    internal class PropertyGridToolStripAccessibleObject : ToolStrip.ToolStripAccessibleObject
-    {
-        private readonly PropertyGrid _parentPropertyGrid;
-
-        /// <summary>
-        ///  Constructs new instance of PropertyGridToolStripAccessibleObject
-        /// </summary>
-        /// <param name="owningPropertyGridToolStrip">The PropertyGridToolStrip owning control.</param>
-        /// <param name="parentPropertyGrid">The parent PropertyGrid control.</param>
-        public PropertyGridToolStripAccessibleObject(PropertyGridToolStrip owningPropertyGridToolStrip, PropertyGrid parentPropertyGrid) : base(owningPropertyGridToolStrip)
-        {
-            _parentPropertyGrid = parentPropertyGrid;
-        }
-
-        /// <summary>
-        ///  Request to return the element in the specified direction.
-        /// </summary>
-        /// <param name="direction">Indicates the direction in which to navigate.</param>
-        /// <returns>Returns the element in the specified direction.</returns>
-        internal override UiaCore.IRawElementProviderFragment FragmentNavigate(UiaCore.NavigateDirection direction)
-        {
-            if (_parentPropertyGrid.IsHandleCreated &&
-                _parentPropertyGrid.AccessibilityObject is PropertyGridAccessibleObject propertyGridAccessibleObject)
-            {
-                UiaCore.IRawElementProviderFragment navigationTarget = propertyGridAccessibleObject.ChildFragmentNavigate(this, direction);
-                if (navigationTarget != null)
-                {
-                    return navigationTarget;
-                }
-            }
-
-            return base.FragmentNavigate(direction);
-        }
-
-        /// <summary>
-        ///  Request value of specified property from an element.
-        /// </summary>
-        /// <param name="propertyID">Identifier indicating the property to return</param>
-        /// <returns>Returns a ValInfo indicating whether the element supports this property, or has no value for it.</returns>
-        internal override object GetPropertyValue(UiaCore.UIA propertyID)
-            => propertyID switch
-            {
-                UiaCore.UIA.ControlTypePropertyId => UiaCore.UIA.ToolBarControlTypeId,
-                UiaCore.UIA.NamePropertyId => Name,
-                _ => base.GetPropertyValue(propertyID)
-            };
-
-        public override string Name
-        {
-            get
-            {
-                string name = Owner?.AccessibleName;
-                if (name != null)
-                {
-                    return name;
-                }
-
-                return _parentPropertyGrid?.AccessibilityObject.Name;
-            }
         }
     }
 }
