+++ src/System.Windows.Forms/tests/UnitTests/System/Windows/Forms/PropertyGridInternal/DetailsButton.DetailsButtonAccessibleObjectTests.cs
@@ -7,7 +7,7 @@ using static Interop;
 
 namespace System.Windows.Forms.PropertyGridInternal.Tests
 {
-    public class DetailsButtonAccessibleObjectTests : IClassFixture<ThreadExceptionFixture>
+    public class DetailsButton_DetailsButtonAccessibleObjectTests : IClassFixture<ThreadExceptionFixture>
     {
         [WinFormsFact]
         public void DetailsButtonAccessibleObject_Ctor_Default()
